refactor: ContactForm component is moved to a seperate file
